+++ scenes/board/logic/state.gd
@@ -113,7 +113,10 @@ func is_current_player_ai():
 	return self.get_current_param("type") == self.PLAYER_AI
 
 func is_player_human(side):
-	return self.players[self.get_player_id_by_side(side)]["type"] == self.PLAYER_HUMAN
+	var player_id = self.get_player_id_by_side(side)
+	if player_id < 0:
+		return false
+	return self.players[player_id]["type"] == self.PLAYER_HUMAN
 
 func is_current_player_alive():
 	return self.get_current_param("alive")