+++ scenes/board_multiplayer/board_multiplayer.gd
@@ -114,6 +114,8 @@ func _can_current_player_perform_actions():
 
 
 func _can_broadcast_moves():
+	if multiplayer.multiplayer_peer == null:
+		return false
 	return _can_current_player_perform_actions() or (self.multiplayer.is_server() and self.state.are_all_peers_present() and self.state.is_current_player_ai())
 
 
@@ -286,3 +288,8 @@ func _notify_player_reconnected():
 	self.all_players_loaded = self.state.are_all_peers_present()
 	_manage_cinematic_bars()
 	_manage_ai_start()
+
+
+func _timer_end_turn() -> void:
+	if _can_broadcast_moves():
+		_end_turn()