+++ docs/CHANGELOG.md
@@ -1,13 +1,24 @@
 # Tanks of Freedom 3D
 ## Changelog
 
+
+### 1.0.7
+- Multiplayer host can now kick players from lobby
+- Added configurable turn time limit to multiplayer games
+- Added configurable turns limit to multiplayer games
+- Added points summary in multiplayer games in case of a turn limit draw
+- Added Online join code copy button
+- Added tilt-shift blur toggle to Settings
+- Moved UI elements in map picker and load/save interface
+
+
 ### 1.0.6
 - Added spectators to multiplayer matches
 - Unit is snapped into position before attacking to prevent misalignment
 - Online registration is now stored on per-domain basis
 - Fixed screen edge camera pan getting stuck if camera is dragged while panning
 - Fixed multiplayer settings domain being converted to a number value
-- Fixed a reggresion introduced in 1.0.3 that prevented players from properly reconnecting to a multiplayer match
+- Fixed a reggresion introduced in 1.0.5 that prevented players from properly reconnecting to a multiplayer match
 
 
 ### 1.0.5